node.rs test cleanup
